status: unbreak the (deprecated) --no-lock-index option
When core Git introduced the --no-optional-locks option in v2.15.0, it
knowingly implemented a different solution to the same problem, causing
merge conflicts and requiring an upgrade to Team Explorer.

To lessen the impact of that, I tried to handle --no-lock-index by
triggering the same code as --no-optional-locks, with a big fat warning
that --no-lock-index is deprecated.

In the last-minute rebasing frenzy leading up to Git for Windows
v2.15.0, however, I managed to forget about the validation of that code
(and I also wanted to update the commit message, which I also forgot).

And of course my patch broke the feature, by using the `no_lock_index`
variable before the command-line options were parsed :-(

This commit fixes it (and also reverts another patch that is no longer
necessary, which persisted the --no-lock-index flag to spawned
processes). It is a combination of two separate fixup commits that
already made it into Git for Windows' `master` branch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>