gvfs: introduce placeholder option
An earlier GVFS Git version limited the commit slob to 1, to avoid
fetching commit objects unnecessarily. However, we now have all the
commit objects locally anyway (we figured out that it is cheaper to
clone all commit and tree objects, they are not really large).

To maintain backwards compatibility, we leave this bit unused for now.

Signed-off-by: Kevin Willford <kewillf@microsoft.com>