gvfs: use a hashmap to back gitignore/sparse checkout
The always_exclude mechanism is used both to implement the excludes
defined in, say, .gitignore, and the sparse checkout feature.

Due to its use of wildcards, any sizable list of entries will slow down
every path-related operation dramatically.

This is particularly noticeable with GVFS, where the sparse-checkout
file is populated and updated frequently by the backend.

In GVFS' case, the use of wildcards is limited to same-directory
wildcards (i.e. something of the form `/this/directory/*`), though,
therefore we can speed up everything dramatically by using exact matches
backed by a hashmap.

Note: the use of a hashmap disallows more free-form sparse-checkout
entries such as `*.txt`. Using a hashmap restricts us to only allow
exact file names, same-directory, and negative entries in the
sparse-checkout file, but with GVFS that is all we need.

This patch introduces the option to make it so.

Signed-off-by: Eric Mecklenburg <ermeckle@microsoft.com>