fixup! sparse-checkout: update files with a modify/delete conflict
This is to make sure that the modified file will get placed on disk for
the modified/deleted merge conflict scenario when using the
sparse-checkout and the file might not be there because of the
skip-worktree bit.

Unfortunately, this patch was not ported correctly onto Git for Windows
2.13.0. This fixup fixes that.

Also fixed the test that should have caught this issue.

Signed-off-by: Kevin Willford <kewillf@microsoft.com>