squash! Hydrate missing loose objects in check_and_freshen_local
TODO-WHEN-SQUASHING: change "check_and_freshen_local" to
"check_and_freshen".

This commit changes the read object hook to handle writing to
alternates.

The read object hook was only handling when it wrote to the local
git objects store (.git/objects).  If it is configured to write to
an alternates location it will always download the objects because
the objects will never be in the local objects directory which is
the only location checked before calling the read object hook to
download.

This change moves where the read object hook is called to after
both the local and nonlocal objects directory are checked. This
makes sure the object isn't in the local or alternates before
attempting to download the object.  After the download, attempt to
get the object from both the local and alternates so that whether
the object is downloaded to the local or the nonlocal objects directory
it will be found.

Signed-off-by: Kevin Willford <kewillf@microsoft.com>