midx: teach git midx --write --update-head
When looking for the "canonical" midx file in a pack directory, find
the file midx-[oid].midx where 'oid' is given by the contents of the
midx-head file in the pack directory.

'git midx --write --update-head' writes a new midx file and updates
midx-head to contain the oid of the new midx file.

Update t/t5317-midx.sh to verify midx-head is updated properly.