graph: load trees lazily from commit-graph file
When walking many commits in the commit-graph file, we may not need all
of the root trees. Delay this load until a read is requested. Use
`get_commit_tree()` and `get_commit_tree_oid()` methods for all
consumers.

This change gets a 20% speedup on `git log --graph --oneline -1000` for
the Linux repo and about 10% speedup for the same command in the OS
repo.

Signed-off-by: Stolee, Derrick <dstolee@microsoft.com>