fixup! commit.c: use generation number to stop merge-base walks
We noticed that there was a regression in `paint_down_to_common()` due
to missing initial conditions of the `min_nonstale_gen`. This probably
didn't show up in my tests because I was using branches that had
undefined generation number (they were not in the graph) but this
doesn't work if all of the branches are in the graph.

Since `paint_down_to_common()` doesn't have unit-level tests, the error
was not caught. The `merge-base` calculations perform a graph walk after
`paint_down_to_common()` that was taking longer than usual, but still
giving correct values.

Signed-off-by: Stolee, Derrick <dstolee@microsoft.com>