merge: load config before loading HEAD
When the commit-graph feature is enabled, it is important that all
commits that exist in the graph are loaded from the graph and not any-
where else. This is especially important when using generation numbers
to make reachability decisions. If the commit is not loaded from the
graph, but one of its descendants is loaded from the graph, then the
generations are swapped.

One case where this is a problem is during a fast-forward merge. In
builtin/merge.c, the config settings are loaded after looking up the HEAD
commit. This means the core_commit_graph setting is not set, so it skips
the commit-graph file.

Reorder the config load to be before the HEAD commit load.

Also create a test in t5318-commit-graph.sh that demonstrates this bug,
and prevents a regression.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>