midx: harden against large offset problems
If a pack-file has an offset larger than 32 bits, we should
write a large offset chunk to the multi-pack-index. This logic
is separated from the place where we actually write offsets.

Add defense in-depth by adding a BUG() statement if we ever see
a large offset but we are not in the mode where we should write
large offsets.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>