dir: adjust for previous merge
The merge of the `submodule-fix-loose` branch introduced a change in the
signature of the `verify_path()` function, which we use in `dir.c`.
Let's just pass 0 as mode, as we're talking about untracked files here,
and we do not care to prevent untracked files from being symlinks, even
if they have a funny name like `.gitmodules`.

This fix was kept out intentionally from the merge commit (which would
have otherwise been an "evil merge", i.e. a merge that introduces more
changes than the merged branches), to make reviewing easier.

Bisectability is still maintained, as the merge commit can be easily
`git bisect skip`ed.