Update the virtualfilesystem support
We now specify that it needs to be run from the root of the git work
tree. This enables the hook to be found even if the current working
directory is not the root of the repo (like when running 'git diff' with
Beyond Compare configured as the diff tool.

Also simpify how argv[] parameter is created.

Signed-off-by: Ben Peart <benpeart@microsoft.com>