fixup! packfile.c: create prepare_packed_git_internal(midx)
Turns out that the current code causes a segmentation fault on a GVFS
clone on macOS: There was a bad memory assignment in the midx code
around an empty midx that doesn't fail on any other platform (and only
on the installed version, perhaps due to extra optimization switches).

Noticed by William Baker.

Signed-off-by: Stolee, Derrick <dstolee@microsoft.com>