fixup! status: add status serialization mechanism
Since 73ba5d78b4 (roll wt_status_state into wt_status and populate in
the collect phase, 2018-09-30), wt_longstatus_print() no longer sets the
state, and hence does not realize when a merge is in progress. This
triggered a test failure in t7524 that specifically tested for "You have
unmerged paths" in the long status.