midx: verify: group objects by packfile to speed up object verification
Sort the set of objects by packfile so that only one packfile needs
to be open at a time.

This is a performance improvement.  Previously, objects were
verified in OID order.  This essentially requires all packfiles
to be open at the same time.  If the number of packfiles exceeds
the open file limit, packfiles would be closed and re-opened
many times.

Signed-off-by: Jeff Hostetler <jeffhost@microsoft.com>