ci: use git-sdk-64-minimal
This is a (partial) backport of the improvements of the Azure Pipeline
made in Git for Windows v2.21.0.

The major benefit is that it makes the task to initialize the Git for
Windows SDK *a lot* faster.

It would also have worked around a problem where an MSYS2 update breaks
t6500.13, and it will work around similar problems in the future, as the
Azure Pipeline that updates git-sdk-64-minimal will not publish a new
artifact unless Git's test suite (as of upstream's current `master`)
passes.