t5516: relax error message pattern
This test in t5516 is failing some times due to a race condition in
the stderr output. The remote is sending an error line in the middle
of the local client writing an error. This causes the exact line
match specified by the path to be incorrect.

Relax the condition a little for now, and consider a better fix to
the race condition later.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>