sparse-checkout: allow cone-type patterns for full tree
Due to an oversight when handling cone patterns in sparse-checkout,
the full cone patternset (only "/*") would not fill any entries in
the hashsets, and only files at root would be populated.

For cone mode patterns, the first two patterns are _usually_ the
following:

	/*
	!/*/

However, we instead need to handle the case where that second
line does not exist. Due to the MUSTBEDIR and NEGATIVE flags, the
second pattern becomes "/*".

Create a new option in struct pattern_list to store this "full
cone" mode. When we see the "/*" pattern with no flags, we toggle
full_cone to on. If we later see "/*" pattern with MUSTBEDIR and
NEGATIVE flags, then turn full_cone off.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>